requirements.txt
# Core dependencies
neo4j==5.15.0
python-dotenv==1.0.0

# Embeddings and ML
openai==1.12.0
google-generativeai==0.3.2
sentence-transformers==2.3.1
numpy==1.26.3

# API framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
mangum==0.17.0

# Utilities
python-dateutil==2.8.2
hashlib-compat==1.0.1
requests==2.31.0

